// src/app/modules/user/user.interface.ts
export type TFullName = {
  firstName: string;
  lastName: string;
};

export type TAddress = {
  street: string;
  city: string;
  country: string;
};

export type TOrder = {
  productName: string;
  price: number;
  quantity: 2;
};

export type TUser = {
  userId: number;
  username: string;
  password: string;
  fullName: TFullName;
  age: number;
  email: string;
  istActive: boolean;
  hobbies: string[];
  address: TAddress;
  orders: TOrder[];
};